+++ example/bug-fixed/cmake-config-cxflags-quoted/package/repro.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+int repro_value();